requirements.txt
# FastAPI and ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Data processing
numpy
pandas

# HTTP client for GRID API
httpx==0.26.0
aiohttp==3.9.3

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
orjson==3.9.13

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
